requirements.txt
streamlit
pandas
pdf2docx
python-docx
docx2pdf
dataframe-image
fpdf